Refactoring of service layer
