+++ src/main/java/com/metrics/service/ErrorHandler/PathErrorMessage.java
@@ -1,4 +1,4 @@
 package com.metrics.service.ErrorHandler;
 public class PathErrorMessage {
-	public static final String pathMetric = "/metric";
+	public static final String pathMetric = "/metrics";
 }
