+++ src/main/java/com/metrics/service/MetricsServiceImpl.java
@@ -88,7 +88,8 @@ public class MetricsServiceImpl implements MetricsService {
 		MetricsApplication.logger.info("Creating metric object");
 		MetricsCollection metric = new MetricsCollection();
 		MetricsApplication.logger.info("calling data validation method");
-		metric = Functions.CreateMetricRequestToMetricsCollection(Functions.testMetricIntegrity(request, 1));
+		metric = TechnicalValidations
+				.CreateMetricRequestToMetricsCollection(ClientValidations.testMetricIntegrity(request, 1));
 		MetricsApplication.logger.info("data validation test passed, saving new object");
 		metric = orikaMapperFacade.map(request, MetricsCollection.class);
 		MetricsApplication.logger.info("object created succesfully");
@@ -153,10 +154,10 @@ public class MetricsServiceImpl implements MetricsService {
 		for (MetricsCollection metric : metrics) {
 			try {
 
-				if ((Functions.stringToDate(metric.getDate()).after(startDate)
-						|| Functions.stringToDate(metric.getDate()).equals(startDate))
-						&& Functions.stringToDate(metric.getDate()).before(endDate)
-						|| Functions.stringToDate(metric.getDate()).equals(endDate)) {
+				if ((TechnicalValidations.stringToDate(metric.getDate()).after(startDate)
+						|| TechnicalValidations.stringToDate(metric.getDate()).equals(startDate))
+						&& TechnicalValidations.stringToDate(metric.getDate()).before(endDate)
+						|| TechnicalValidations.stringToDate(metric.getDate()).equals(endDate)) {
 					listMetricsFiltredDates.add(metric);
 				}
